refactor: startEngineLoop
